feat: intergrate common feature
